MACOSXPPC: Follow upstream dockerized-bb's update for SDL1.2, minus one OSXPPC PR
I avoided updating the SDL1.2 snapshots until now, because of the
https://github.com/libsdl-org/SDL-1.2/pull/871 changes I don't trust
much (they just look untested).

Now I think it's time to sync our SDL1.2 snapshot again, but I've
chosen to revert this PR. It just looks like it reinstates code
that was disabled for years and years, and it looks like it only
affected OSX 10.3 and below, for which we don't care.

Maybe there's something good to be picked up in this PR, but for
now I'm not aware of a major issue regarding SDL1.2, and I've spent
enough time trying to understand what this old code does.

My findings were added to the top of this patch.

SDL1.2 remains built with /usr/bin/gcc-4.0 because of its reliance
on the Apple-only behavior of the -faltivec flag (otherwise, we either
loose Altivec acceleration code, or G3 compatibility).