MACOSXPPC: Still no Fluidlite support in 2.9.0 because of issues on big-endian
When playing Blazing Dragons, I get a wrongly byte-swapped 'size'
chunk values, and bogus audio, unless I remove some byte-swapping
code from Fluidlite itself.

Since I can't say whether the issue is in Fluidlite itself or in
ScummVM or in the dragons engine, I chose not to enable the Fluidlite
feature for the official 2.9.0 build. I've spend a couple of days on
trying to understand this better, but I'll need help.