MACOSXPPC: Sync with latest upstream dockerized-bb changes
Sync with e4a5dd0d684b9a91ec23116f10bcf83d18877d69 upstream.

This brings the Debian 13 package version upgrade that was done there,
it also syncs with some custom updates it has, and then we have some
custom changes:

* libpng is a bit newer
* libtheora is still using the old 1.1.1 release, with some newer
  Debian patches for it

for reasons explained in their build scripts.

This is what's should be used for the OSXPPC builds of ScummVM 3.0.0.

UNTESTED for now.